Fix and organize batch burn tests
